requirements.txt
# MongoDB MCP Client Dependencies
mcp>=1.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
rich>=13.7.0

